Go back to .NET 6
The upgrade broke two things that matter, and the second is bad enough to be
worth undoing the whole move rather than patching around it.

Unauthenticated page loads answered 401 instead of redirecting, so anyone on
Forms authentication got an empty error page with no way to log in. That one
had a fix.

The second is that request bodies stopped binding on any action that relies on
inference rather than an explicit [FromBody] - 31 of the 55 POST and PUT
actions. Commands arrived with no name, so the refresh and search buttons did
nothing, and grabbing a release from interactive search failed on an empty
guid and a zero indexer id. That is most of the write surface of the API, and
finding the cause is a bigger job than it is worth doing against a live
install.

The branch is still on the remote with all of it intact - the retarget, the
obsolete API fixes, the vulnerable package bumps and the array Contains fix
the where builders needed. Worth picking up again as its own piece of work,
starting from why binding inference stopped.

+++ src/NzbDrone.Core/Download/Clients/Transmission/TransmissionBase.cs
@@ -78,9 +78,7 @@ namespace NzbDrone.Core.Download.Clients.Transmission
                 {
                     try
                     {
-                        // Cast to double deliberately: the long overload added in .NET 9 throws
-                        // ArgumentOutOfRangeException instead of the OverflowException this falls back on.
-                        item.RemainingTime = TimeSpan.FromSeconds((double)torrent.Eta);
+                        item.RemainingTime = TimeSpan.FromSeconds(torrent.Eta);
                     }
                     catch (OverflowException)
                     {
